+++ packages/react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 5.0.0-beta.32
+
+### Patch Changes
+
+- Updated dependencies [[`afec17953`](https://github.com/clerk/javascript/commit/afec17953d1ae4ba39ee73e4383757694375524d)]:
+  - @clerk/types@4.0.0-beta.22
+
 ## 5.0.0-beta.31
 
 ### Patch Changes